+++ modules/raft/src/main/java/org/apache/ignite/raft/jraft/core/NodeImpl.java
@@ -1264,7 +1264,6 @@ public class NodeImpl implements Node, RaftServerService {
                 false,
                 false,
                 null
-                //opts.getRaftMetrics().disruptorMetrics("JRaft-FSMCaller-Disruptor")
             ));
         }
