+++ modules/table/build.gradle
@@ -109,6 +109,7 @@ dependencies {
     testFixturesImplementation(testFixtures(project(':ignite-placement-driver-api')))
     testFixturesImplementation(testFixtures(project(':ignite-low-watermark')))
     testFixturesImplementation(testFixtures(project(':ignite-failure-handler')))
+    testFixturesImplementation(testFixtures(project(':ignite-metrics')))
     testFixturesImplementation libs.jetbrains.annotations
     testFixturesImplementation libs.fastutil.core
     testFixturesImplementation libs.mockito.core
@@ -155,6 +156,7 @@ dependencies {
     integrationTestImplementation(testFixtures(project(':ignite-runner')))
     integrationTestImplementation(testFixtures(project(':ignite-low-watermark')))
     integrationTestImplementation(testFixtures(project(':ignite-failure-handler')))
+    integrationTestImplementation(testFixtures(project(':ignite-metrics')))
     integrationTestImplementation libs.fastutil.core
     integrationTestImplementation libs.jetbrains.annotations
     integrationTestImplementation libs.calcite.core
